change!: move the hashing API to `gix_hash`
This will soon not be dependent on any feature configuration. I
wasn’t sure the ideal crate to put this in, but after checking
reverse dependencies on crates.io, it seems like there’s essentially
no user of `gix-hash` that wouldn’t be pulling in a hashing
implementation anyway, so I think this is a fine and logical place
for it to be.

I named the module `gix_hash::hasher` since `gix_hash::hash` seemed
like it would be confusing. This does mean that there is a function
and module with the same name, which is permitted but perhaps a
little strange.

Everything is re‐exported directly other than
`gix_features::hash::Write`, which moves along with the I/O
convenience functions into a new public submodule and becomes
`gix_hash::hasher::io::Write`, as that seems like a clearer name
to me, being akin to the `gix_hash::hasher` function but as an
`std::io::Write` wrapper.