change!: make the hashing API return `ObjectId`
The vast majority of users of this API immediately converted the result
to `ObjectId`, so this eliminates a lot of cruft across the tree. It
also has a nicer `Debug` and `Display` instance than `[u8; 20]`,
and should theoretically make supporting the hash function transition
easier, although I suspect further API changes will be required for
that anyway.

I wasn’t sure whether this would be a good change, as not every
digest identifies an entry in the Git object database, but even many
of the existing uses for non‐object digests across the tree used the
`ObjectId` API anyway. Perhaps it would be best to have a separate
non‐alias `Digest` type that `ObjectId` wraps, but this seems like
the pragmatic choice for now that sticks with current practice.