change!: make hashing fallible to prepare for collision detection
This seems better than killing the process as Git does, since
we’re in a library context and it would be bad if you could
perform denial‐of‐service attacks on a server by sending it hash
collisions. (Although there are probably cheaper ways to mount a
denial‐of‐service attack.)

It does mean a lot of churn across the tree, but the underlying
breaking change is to a low‐level API, and the change is usually
just an adjustment to variants of an existing error type, so I expect
that most downstream users will require little to no adaption for
this change.