fix!: detect SHA‐1 collision attacks
Fix [GHSA-2frx-2596-x5r6].

[GHSA-2frx-2596-x5r6]: https://github.com/GitoxideLabs/gitoxide/security/advisories/GHSA-2frx-2596-x5r6

This uses the `sha1collisiondetection` crate from the Sequoia PGP
project, based on a modified C2Rust translation of the library
used by Git. The `sha1-checked` crate was also considered, but
`sha1collisiondetection` measures about 1.13× its performance in
my testing.

It uses a pure Rust implementation, with no SIMD or assembly
code. Raw hashing is about 0.26× to 0.64× the speed of the
previous implementation, depending on the feature configuration
and whether the CPU supports hardware‐accelerated hashing. (The
more portable assembly in `sha1-asm` that doesn’t require the SHA
instruction set doesn’t seem to speed things up that much; in fact,
`sha1_smol` somehow regularly beats the assembly code used by `sha1`
on my i9‐9880H MacBook Pro! Presumably this is why that path was
removed in newer versions of the `sha1` crate.)

That’s very unfortunate, but the [SHAttered] attack demonstrated a
collision back in 2017, and the 2020 [SHA‐1 is a Shambles] attack
demonstrated a practical chosen‐prefix collision that broke the use
of SHA‐1 in OpenPGP, costing $75k to perform, with an estimate of
$45k to replicate at the time of publication and $11k for a classical
collision.

Given the increase in GPU performance and production since then,
that puts the Git object format squarely at risk. Git mitigated this
attack in 2017; the algorithm is fairly general and detects all the
existing public collisions. My understanding is that an entirely new
cryptanalytic approach would be required to develop a collision attack
for SHA‐1 that would not be detected with very high probability.

[SHAttered]: https://shattered.io/
[SHA‐1 is a Shambles]: https://sha-mbles.github.io/

I believe that the speed penalty could be mitigated, although not
fully eliminated, by implementing a version of the hardened SHA‐1
function that makes use of SIMD. For instance, the assembly code used
by `openssl speed sha1` on my i9‐9880H measures around 830 MiB/s,
compared to the winning 580 MiB/s of `sha1_smol`; adding collision
detection support to that would surely incur a performance penalty,
but it is likely that it could be much more competitive with
the performance before this commit than the 350 MiB/s I get with
`sha1collisiondetection`. I haven’t been able to find any existing
work on this; it seems that more or less everyone just uses the
original C library that Git does, presumably because nothing except
Git and OpenPGP is still relying on SHA‐1 anyway…

The performance will never compete with the >2 GiB/s that can
be achieved with the x86 SHA instruction set extension, as the
`SHA1RNDS4` instruction sadly runs four rounds at a time while the
collision detection algorithm requires checks after every round,
but I believe SIMD would still offer a significant improvement,
and the AArch64 extension seems like it may be more flexible.

I know that these days the Git codebase has an additional faster
unsafe API without these checks that it tries to carefully use only
for operations that do not depend on hashing results for correctness
or safety. I personally believe that’s not a terribly good idea,
as it seems easy to misuse in a case where correctness actually does
matter, but maybe that’s just my Rust safety bias talking. I think
it would be better to focus on improving the performance of the safer
algorithm, as I think that many of the operations where the performance
penalty is the most painful are dealing with untrusted input anyway.

The `Hasher` struct gets a lot bigger; I don’t know if this is
an issue or not, but if it is, it could potentially be boxed at the
expense of a further performance penalty. The `sha1-checked` crate
has a smaller type, but is a little bit slower in my testing.

Closes: #585